// src/redux/painting/types.ts
export type PaintingItem = {
  id: string;
  title: string;
  price: number;
  imageUrl: string;
  sizes: number[];
  types: number[];
  rating: number;
};

export enum Status {
  LOADING = "loading",
  SUCCESS = "success",
  ERROR = "error",
}

export interface PaintingSliceState {
  items: PaintingItem[];
  status: Status;
}